requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
asyncer==0.0.2
attrs==24.2.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
build==1.2.1
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
chainlit==1.1.402
chardet==5.2.0
charset-normalizer==3.3.2
chevron==0.14.0
chroma-hnswlib==0.7.6
chromadb==0.5.5
click==8.1.7
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
dataclasses-json==0.5.14
debugpy==1.8.5
decorator==5.1.1
deepdiff==7.0.1
Deprecated==1.2.14
distro==1.9.0
effdet==0.4.1
emoji==2.12.1
et-xmlfile==1.1.0
executing==2.0.1
fastapi==0.110.3
filelock==3.15.4
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
google-api-core==2.19.1
google-auth==2.33.0
google-cloud-vision==3.7.4
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.65.4
grpcio-status==1.65.4
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.5
humanfriendly==10.0
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.0
iopath==0.1.10
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
kubernetes==30.1.0
langchain==0.2.12
langchain-chroma==0.1.2
langchain-community==0.2.11
langchain-core==0.2.28
langchain-openai==0.1.20
langchain-text-splitters==0.2.2
langdetect==1.0.9
langsmith==0.1.98
layoutparser==0.3.4
Lazify==0.4.0
literalai==0.0.607
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.1.post1
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
omegaconf==2.3.0
onnx==1.16.2
onnxruntime==1.18.1
openai==1.40.0
opencv-python==4.10.0.84
openpyxl==3.1.5
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
ordered-set==4.1.0
orjson==3.10.6
overrides==7.7.0
packaging==23.2
pandas==2.2.2
parso==0.8.4
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.2
pexpect==4.9.0
pikepdf==9.1.0
pillow==10.4.0
pillow_heif==0.18.0
platformdirs==4.2.2
portalocker==2.10.1
posthog==3.5.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.4
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.8
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.9.0
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pypandoc==1.13
pyparsing==3.1.2
pypdf==4.3.1
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.1.0
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-engineio==4.9.1
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.9
python-oxmsg==0.0.1
python-pptx==0.6.23
python-socketio==5.11.3
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.1.0
rapidfuzz==3.9.6
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rsa==4.9
safetensors==0.4.4
scipy==1.14.0
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.32
stack-data==0.6.3
starlette==0.37.2
sympy==1.13.1
syncer==2.0.3
tabulate==0.9.0
tenacity==8.5.0
tiktoken==0.7.0
timm==1.0.8
tokenizers==0.19.1
tomli==2.0.1
torch==2.2.2
torchvision==0.17.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.0
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
unstructured==0.15.1
unstructured-client==0.25.4
unstructured-inference==0.7.36
unstructured.pytesseract==0.3.12
uptrace==1.26.0
urllib3==2.2.2
uvicorn==0.25.0
uvloop==0.19.0
watchfiles==0.20.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==12.0
wrapt==1.16.0
wsproto==1.2.0
xlrd==2.0.1
XlsxWriter==3.2.0
yarl==1.9.4
zipp==3.19.2
